+++ crates/trycp_server/src/main.rs
@@ -67,7 +67,7 @@ async fn main() -> Result<(), Error> {
             long = "lair-shim",
             help = "Server code for a lair shim i.e a replacement for lair-keystore",
         )]
-        lair_shim: Option<u16>,
+        lair_shim: Option<u64>,
     }
     let args = Cli::from_args();
 
@@ -208,7 +208,7 @@ enum ConnectionError {
     },
 }
 
-async fn ws_connection(stream: TcpStream, lair_shim: Option<u16>) -> Result<(), ConnectionError> {
+async fn ws_connection(stream: TcpStream, lair_shim: Option<u64>) -> Result<(), ConnectionError> {
     let ws_stream = tokio_tungstenite::accept_async(stream)
         .await
         .context(Handshake)?;
@@ -236,7 +236,7 @@ async fn ws_connection(stream: TcpStream, lair_shim: Option<u16>) -> Result<(), 
 async fn ws_message(
     message_res: Result<Message, tungstenite::Error>,
     ws_write: Arc<futures::lock::Mutex<WsResponseWriter>>,
-    lair_shim: Option<u16>,
+    lair_shim: Option<u64>,
 ) -> Result<Option<Message>, ConnectionError> {
     let message = message_res.context(ReadRequest)?;
